requirements.txt
beautifulsoup4
lxml
requests
markdownify
python-dotenv
markdown
jinja2
pyyaml
flask
flask_limiter
bleach
pillow
pyheif